// src/main/webapp/src/api/models/Settings.ts
export interface Settings {
    /**
     * 
     * @type {Range}
     * @memberof Settings
     */
    nbProParticipants?: Range;
    /**
     * 
     * @type {Range}
     * @memberof Settings
     */
    numberOfAssignmentsForAProfessional?: Range;
    /**
     * 
     * @type {Range}
     * @memberof Settings
     */
    nbNonProParticipants?: Range;
    /**
     * 
     * @type {Range}
     * @memberof Settings
     */
    numberOfAssignmentsForANonProfessional?: Range;
    /**
     * 
     * @type {Range}
     * @memberof Settings
     */
    nbExternalParticipants?: Range;
    /**
     * 
     * @type {Range}
     * @memberof Settings
     */
    numberOfAssignmentsForAnExternal?: Range;
    /**
     * 
     * @type {number}
     * @memberof Settings
     */
    nbRotationsToReinspect?: number;
    /**
     * 
     * @type {number}
     * @memberof Settings
     */
    nbInspectorsFollowingUp?: number;
    /**
     * 
     * @type {DistanceMatrix}
     * @memberof Settings
     */
    distanceMatrix?: DistanceMatrix;
    /**
     * 
     * @type {Range}
     * @memberof Settings
     */
    travellingDistanceRange?: Range;
    /**
     * 
     * @type {boolean}
     * @memberof Settings
     */
    useAvailability?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof Settings
     */
    shuffleParticipants?: boolean;
    /**
     * 
     * @type {Range}
     * @memberof Settings
     */
    committeeMeetingSize?: Range;
}