// src/main/webapp/src/api/models/SolverOptions.ts
/* tslint:disable */
/* eslint-disable */
/**
 * dicoop API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.13.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { Person } from './Person';
import {
    PersonFromJSON,
    PersonFromJSONTyped,
    PersonToJSON,
    PersonToJSONTyped,
} from './Person';
import type { Settings } from './Settings';
import {
    SettingsFromJSON,
    SettingsFromJSONTyped,
    SettingsToJSON,
    SettingsToJSONTyped,
} from './Settings';

/**
 * 
 * @export
 * @interface SolverOptions
 */
export interface SolverOptions {
    /**
     * 
     * @type {Settings}
     * @memberof SolverOptions
     */
    settings?: Settings;
    /**
     * 
     * @type {Array<Person>}
     * @memberof SolverOptions
     */
    participants?: Array<Person>;
}

/**
 * Check if a given object implements the SolverOptions interface.
 */
export function instanceOfSolverOptions(value: object): value is SolverOptions {
    return true;
}

export function SolverOptionsFromJSON(json: any): SolverOptions {
    return SolverOptionsFromJSONTyped(json, false);
}

export function SolverOptionsFromJSONTyped(json: any, ignoreDiscriminator: boolean): SolverOptions {
    if (json == null) {
        return json;
    }
    return {
        
        'settings': json['settings'] == null ? undefined : SettingsFromJSON(json['settings']),
        'participants': json['participants'] == null ? undefined : ((json['participants'] as Array<any>).map(PersonFromJSON)),
    };
}

export function SolverOptionsToJSON(json: any): SolverOptions {
    return SolverOptionsToJSONTyped(json, false);
}

export function SolverOptionsToJSONTyped(value?: SolverOptions | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'settings': SettingsToJSON(value['settings']),
        'participants': value['participants'] == null ? undefined : ((value['participants'] as Array<any>).map(PersonToJSON)),
    };
}